add the first sent flag
