+++ extensions/applicationinsights-analytics-js/src/JavaScriptSDK/AnalyticsPlugin.ts
@@ -285,22 +285,6 @@ export class AnalyticsPlugin extends BaseTelemetryPlugin implements IAppInsights
                 if (doc) {
                     pageView.refUri = pageView.refUri === undefined ? doc.referrer : pageView.refUri;
                 }
-
-                let perf = getPerformance();
-                // Access the performance timing object
-                const navigationEntries = (perf && perf.getEntriesByType && perf.getEntriesByType("navigation"));
-
-                // Edge Case the navigation Entries may return an empty array and the timeOrigin is not supported on IE
-                if (navigationEntries && navigationEntries[0] && !isUndefined(perf.timeOrigin)) {
-                    // Get the value of loadEventStart
-                    const loadEventStart = (navigationEntries[0] as PerformanceNavigationTiming).loadEventStart;
-                    pageView.startTime =  new Date(perf.timeOrigin + loadEventStart);
-                } else {
-                    // calculate the start time manually
-                    let duration = ((properties || pageView.properties || {}).duration || 0);
-                    pageView.startTime = new Date(new Date().getTime() - duration);
-                }
-
                 let telemetryItem = createTelemetryItem<IPageViewTelemetryInternal>(
                     pageView,
                     PageView.dataType,
