+++ shared/AppInsightsCommon/src/Interfaces/IPageViewTelemetry.ts
@@ -62,5 +62,4 @@ export interface IPageViewTelemetryInternal extends IPageViewTelemetry {
      * Version of the part B schema, todo: set this value in trackpageView
      */
     ver?: string;
-    startTime?: Date;
 }
